Fix code typos and module call for Template fairing
The call to diesel to create users and posts should be done with 'diesel
migration ...' and not 'diesel migrations...', ie. no 's'. In the index
function, there should be no semi-colon at the end of the Template
render line since it is the return value of the function. When attaching
the Template fairing, the fairing module should be imported with two
colons after 'Template'. The base template file code is missing a '/' at
 the end of the first paragraph.